requirements.txt
# Streamlit
streamlit>=1.40.0
streamlit-extras>=0.4.0

# LangChain
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-community>=0.2.0
langgraph>=0.2.0

# AI
openai>=1.57.0

# 仓库分析
GitPython>=3.1.41
PyGithub>=2.1.1

# 数据处理
pydantic>=2.6.0
python-dotenv>=1.0.0
httpx>=0.26.0

# 代码高亮
pygments>=2.17.2

# Markdown 渲染
markdown>=3.5.2